rec-api/requirements.txt
alembic==1.4.3
aniso8601==8.0.0
attrs==20.2.0
bcrypt==3.2.0
cffi==1.14.3
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.5.3
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.0
Flask-Caching==1.10.1
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
Mako==1.1.3
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==1.7.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.20
Werkzeug==0.16.0
zipp==3.4.0
flask-mongoengine==1.0.0
hvac==0.10.6
pymongo==3.11.2
mongoengine~=0.22.1
pytest==6.1.2
setuptools==49.2.0
wordninja==2.0.0
names_dataset==1.9.1
requests==2.23.0
PyMySQL==1.0.2
cython==0.29.22
numpy==1.17.3
scipy==1.3.2
xgboost==1.3.3
scikit-learn==0.24.1
gensim==3.8.3
pandas==1.2.3
nltk==3.5
DateTime==4.3
simplejson==3.17.2
pyarrow==4.0.0
mongomock==3.23.0
pytest-cov==2.12.1
redis~=3.5.3
fakeredis~=1.5.0
google-cloud-pubsub==2.8.0
psutil==5.8.0
google~=3.0.0
pycryptodome~=3.14
redis-py-cluster~=2.1.3
apache-skywalking==0.7.0
